+++ rules/plugin-chai-expect.js
@@ -3,6 +3,7 @@
 module.exports = {
 	rules: {
 		'chai-expect/no-inner-compare': 'error',
+		'chai-expect/no-inner-literal': 'error',
 		'chai-expect/missing-assertion': 'error',
 		'chai-expect/terminating-properties': 'error',
 	},