+++ rules/plugin-jsdoc.js
@@ -2,6 +2,7 @@
 
 module.exports = {
 	rules: {
+		'jsdoc/check-access': 'error',
 		'jsdoc/check-alignment': 'error',
 		'jsdoc/check-examples': 'error',
 		'jsdoc/check-indentation': 'error',
@@ -9,6 +10,8 @@ module.exports = {
 		'jsdoc/check-syntax': 'error',
 		'jsdoc/check-tag-names': 'error',
 		'jsdoc/check-types': 'error',
+		'jsdoc/check-values': 'off',
+		'jsdoc/empty-tags': 'error',
 		'jsdoc/implements-on-classes': 'error',
 		'jsdoc/newline-after-description': 'error',
 		'jsdoc/match-description': 'error',